Revert "Core/Entities: allow creatures/gameobjects to have 4 byte phaseMask in DB"
This reverts commit a256ec037f18a8f5b0263a265d3aaf19e2878009.

+++ src/server/game/Entities/Creature/Creature.h
@@ -237,7 +237,7 @@ struct CreatureData
     explicit CreatureData() : dbData(true) {}
     uint32 id;                                              // entry in creature_template
     uint16 mapid;
-    uint32 phaseMask;
+    uint16 phaseMask;
     uint32 displayid;
     int32 equipmentId;
     float posX;
