+++ src/server/game/Entities/GameObject/GameObject.h
@@ -565,7 +565,7 @@ struct GameObjectData
     explicit GameObjectData() : dbData(true) {}
     uint32 id;                                              // entry in gamobject_template
     uint16 mapid;
-    uint32 phaseMask;
+    uint16 phaseMask;
     float posX;
     float posY;
     float posZ;
